Script version 5 release
Version 5 fixes the "out" configuration option. Briefly tested to be working with flying colors.

Small fix to argument parsing, where it would incorrectly set value if there was a column in the value area of "key:value" when passing command line arguments.

For subsequent script updates implemented one-time pass of arguments to continue with conversion after an update. This was tested to be working once, using a customized environment so hopefully it works in production. Will test immediately after.

Perhaps I forgot at some point, but I noticed cmd passes the variable environment when utilizing the start command. So hypothetically it should not be necessary to pass arguments after an update. It's just that for now, I wanted it to parse the user section properly anew. Both methods work and have their pros and cons. In an ideal world, I would probably just use the passed environment, since I trust that more in terms of special characters, like quotes.